More ingestion attempts -- broken
Currently broken due to a fundamental issue.. you cannot use the CLOB
client to fetch orders from a user that is not you (or that you have the
creds for).

Wrote a POC watchTrades.ts to watch all the trades -- can filter on the
client side for trades you are interested in.